Rename button and state to finalize
